:recycle: refactor: retirando responsabilidades do arquivo server e passando para app.js #139
